+++ client/src/main/java/client/components/ListComponentCtrl.java
@@ -10,7 +10,9 @@ import java.util.List;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.input.ClipboardContent;
 import javafx.scene.input.Dragboard;
@@ -191,7 +193,13 @@ public class ListComponentCtrl extends VBox {
 
     @FXML
     private void deleteList() {
-        server.deleteCardList(listId);
+        Alert confirmationDialogue = new Alert(Alert.AlertType.CONFIRMATION,
+                "Delete this list ?", ButtonType.YES, ButtonType.NO);
+        confirmationDialogue.showAndWait();
+        if (confirmationDialogue.getResult() == ButtonType.YES) {
+            server.deleteCardList(listId);
+        }
+
     }
 
     @FXML
